+++ libs/langchain-mcp-adapters/src/client.ts
@@ -1,16 +1,7 @@
-import { StdioClientTransport } from "@modelcontextprotocol/sdk/client/stdio.js";
-import {
-  SSEClientTransport,
-  type SseError,
-} from "@modelcontextprotocol/sdk/client/sse.js";
-import {
-  StreamableHTTPClientTransport,
-  type StreamableHTTPError,
-} from "@modelcontextprotocol/sdk/client/streamableHttp.js";
-import type { OAuthClientProvider } from "@modelcontextprotocol/sdk/client/auth.js";
+import { SSEClientTransport, StreamableHTTPClientTransport, SdkHttpError } from "@modelcontextprotocol/client";
+import type { SseError, OAuthClientProvider, LoggingLevel } from "@modelcontextprotocol/client";
+import { StdioClientTransport } from "@modelcontextprotocol/client/stdio";
 import type { DynamicStructuredTool } from "@langchain/core/tools";
-import type { LoggingLevel } from "@modelcontextprotocol/sdk/types.js";
-
 import { z } from "zod/v3";
 import { loadMcpTools } from "./tools.js";
 import { ConnectionManager, type Client } from "./connection.js";
@@ -758,8 +749,11 @@ export class MultiServerMCPClient {
   }
 
   private _getHttpErrorCode(error: unknown): number | undefined {
-    const streamableError = error as StreamableHTTPError | SseError;
-    let { code } = streamableError;
+    const streamableError = error as SdkHttpError | SseError;
+    // v2: SdkHttpError carries the HTTP status on `.status` (its `.code` is an SdkErrorCode enum);
+    // SseError still carries the numeric status on `.code`.
+    let code: number | undefined =
+      streamableError instanceof SdkHttpError ? streamableError.status : streamableError.code;
     // try parsing from error message if code is not set
     if (code == null) {
       const m = streamableError.message.match(/\(HTTP (\d\d\d)\)/);
