--- libs/langchain-mcp-adapters/src/tests/__mocks__/@modelcontextprotocol/sdk/client/sse.ts
@@ -1,18 +1,0 @@
-/* oxlint-disable @typescript-eslint/no-explicit-any */
-import { vi } from "vitest";
-
-const sseClientTransportPrototype = {
-  connect: vi.fn().mockReturnValue(Promise.resolve()),
-  send: vi.fn().mockReturnValue(Promise.resolve()),
-  close: vi.fn().mockReturnValue(Promise.resolve()),
-};
-export const SSEClientTransport: any = vi
-  .fn()
-  .mockImplementation(function mockSSEClientTransport(url, options) {
-    return {
-      ...sseClientTransportPrototype,
-      url,
-      options,
-    };
-  });
-SSEClientTransport.prototype = sseClientTransportPrototype;
