--- libs/langchain-mcp-adapters/src/tests/__mocks__/@modelcontextprotocol/sdk/client/streamableHttp.ts
@@ -1,19 +1,0 @@
-/* oxlint-disable @typescript-eslint/no-explicit-any */
-import { vi } from "vitest";
-
-const streamableHTTPClientTransportPrototype = {
-  connect: vi.fn().mockReturnValue(Promise.resolve()),
-  send: vi.fn().mockReturnValue(Promise.resolve()),
-  close: vi.fn().mockReturnValue(Promise.resolve()),
-};
-export const StreamableHTTPClientTransport: any = vi
-  .fn()
-  .mockImplementation(function mockStreamableHTTPClientTransport(url, options) {
-    return {
-      ...streamableHTTPClientTransportPrototype,
-      url,
-      options,
-    };
-  });
-StreamableHTTPClientTransport.prototype =
-  streamableHTTPClientTransportPrototype;
