+++ libs/langchain-mcp-adapters/src/tests/client.test.ts
@@ -1,8 +1,7 @@
 import { describe, it, expect, beforeEach, afterEach } from "vitest";
 import { Server } from "node:http";
 import { join } from "node:path";
-import type { OAuthClientProvider } from "@modelcontextprotocol/sdk/client/auth.js";
-import type { OAuthTokens } from "@modelcontextprotocol/sdk/shared/auth.js";
+import type { OAuthClientProvider, OAuthTokens } from "@modelcontextprotocol/client";
 import type { ContentBlock } from "@langchain/core/messages";
 import type { ToolCall } from "@langchain/core/messages";
 import type { StructuredToolInterface } from "@langchain/core/tools";
