+++ libs/langchain-mcp-adapters/src/tests/connection.test.ts
@@ -1,27 +1,13 @@
 import { describe, test, expect, beforeEach, vi, type Mock } from "vitest";
-
-import { Client as SDKClient } from "@modelcontextprotocol/sdk/client/index.js";
-import { StdioClientTransport } from "@modelcontextprotocol/sdk/client/stdio.js";
-import { SSEClientTransport } from "@modelcontextprotocol/sdk/client/sse.js";
-import { StreamableHTTPClientTransport } from "@modelcontextprotocol/sdk/client/streamableHttp.js";
-
+import { StdioClientTransport } from "@modelcontextprotocol/client/stdio";
+import { Client as SDKClient, SSEClientTransport, StreamableHTTPClientTransport } from "@modelcontextprotocol/client";
 import { ConnectionManager, type Client } from "../connection.js";
 
-vi.mock(
-  "@modelcontextprotocol/sdk/client/index.js",
-  () => import("./__mocks__/@modelcontextprotocol/sdk/client/index.js")
+vi.mock("@modelcontextprotocol/client", () =>
+  import("./__mocks__/@modelcontextprotocol/client.js")
 );
-vi.mock(
-  "@modelcontextprotocol/sdk/client/stdio.js",
-  () => import("./__mocks__/@modelcontextprotocol/sdk/client/stdio.js")
-);
-vi.mock(
-  "@modelcontextprotocol/sdk/client/sse.js",
-  () => import("./__mocks__/@modelcontextprotocol/sdk/client/sse.js")
-);
-vi.mock(
-  "@modelcontextprotocol/sdk/client/streamableHttp.js",
-  () => import("./__mocks__/@modelcontextprotocol/sdk/client/streamableHttp.js")
+vi.mock("@modelcontextprotocol/client/stdio", () =>
+  import("./__mocks__/@modelcontextprotocol/client/stdio.js")
 );
 
 describe("ConnectionManager", () => {
