+++ libs/langchain-mcp-adapters/src/tests/fixtures/dummy-http-server.ts
@@ -1,15 +1,10 @@
 import { randomUUID } from "node:crypto";
 
 import express, { type Express } from "express";
-
-import {
-  McpServer,
-  ResourceTemplate,
-} from "@modelcontextprotocol/sdk/server/mcp.js";
-import { StreamableHTTPServerTransport } from "@modelcontextprotocol/sdk/server/streamableHttp.js";
-import { SSEServerTransport } from "@modelcontextprotocol/sdk/server/sse.js";
-import { isInitializeRequest } from "@modelcontextprotocol/sdk/types.js";
-import { z } from "zod/v3";
+import { SSEServerTransport } from "@modelcontextprotocol/server-legacy/sse";
+import { NodeStreamableHTTPServerTransport } from "@modelcontextprotocol/node";
+import { McpServer, ResourceTemplate, isInitializeRequest } from "@modelcontextprotocol/server";
+import { z } from "zod";
 
 export function createDummyHttpServer(
   name: string,
@@ -40,213 +35,176 @@ export function createDummyHttpServer(
     // Add tools that can inspect request details
     // oxlint-disable-next-line @typescript-eslint/ban-ts-comment
     // @ts-ignore - this may raise "Type instantiation is excessively deep and possibly infinite.ts(2589)"
-    server.tool(
-      "test_tool",
-      "A test tool that echoes input and request metadata",
-      { input: z.string() },
-      async ({ input }, extra) => {
-        // Logging message
-        await server.server.notification(
-          {
-            method: "notifications/message",
-            params: {
-              level: "info",
-              logger: "test_tool",
-              data: `test_tool invoked with ${input}`,
-            },
-          },
-          { relatedRequestId: extra.requestId }
-        );
+    server.registerTool("test_tool", { description: "A test tool that echoes input and request metadata", inputSchema: z.object({ input: z.string() }) }, async ({ input }, ctx) => {
+                  // Logging message
+                  await server.server.notification(
+                    {
+                      method: "notifications/message",
+                      params: {
+                        level: "info",
+                        logger: "test_tool",
+                        data: `test_tool invoked with ${input}`,
+                      },
+                    },
+                    { relatedRequestId: ctx.mcpReq.id }
+                  );
 
-        // Progress with token if present
-        const progressToken = extra._meta?.progressToken;
-        if (progressToken !== undefined) {
-          const steps = 3;
-          for (let i = 1; i <= steps; i++) {
-            await server.server.notification(
-              {
-                method: "notifications/progress",
-                params: { progress: i, total: steps, progressToken },
-              },
-              { relatedRequestId: extra.requestId }
-            );
-          }
-        }
+                  // Progress with token if present
+                  const progressToken = ctx.mcpReq._meta?.progressToken;
+                  if (progressToken !== undefined) {
+                    const steps = 3;
+                    for (let i = 1; i <= steps; i++) {
+                      await server.server.notification(
+                        {
+                          method: "notifications/progress",
+                          params: { progress: i, total: steps, progressToken },
+                        },
+                        { relatedRequestId: ctx.mcpReq.id }
+                      );
+                    }
+                  }
 
-        return {
-          content: [
-            {
-              type: "text",
-              text: JSON.stringify({
-                input,
-                meta: extra._meta,
-                serverName: name,
-              }),
-            },
-          ],
-        };
-      }
-    );
+                  return {
+                    content: [
+                      {
+                        type: "text",
+                        text: JSON.stringify({
+                          input,
+                          meta: ctx.mcpReq._meta,
+                          serverName: name,
+                        }),
+                      },
+                    ],
+                  };
+                });
 
     // oxlint-disable-next-line @typescript-eslint/ban-ts-comment
     // @ts-ignore - this may raise "Type instantiation is excessively deep and possibly infinite.ts(2589)"
-    server.tool(
-      "sleep_tool",
-      "A test tool that sleeps for the given number of milliseconds before returning",
-      { sleepMsec: z.number().int().positive() },
-      async ({ sleepMsec }) => {
-        await new Promise((resolve) => {
-          setTimeout(resolve, sleepMsec);
-        });
-        return {
-          content: [
-            {
-              type: "text",
-              text: JSON.stringify({
-                message: "done",
-              }),
-            },
-          ],
-        };
-      }
-    );
+    server.registerTool("sleep_tool", { description: "A test tool that sleeps for the given number of milliseconds before returning", inputSchema: z.object({ sleepMsec: z.number().int().positive() }) }, async ({ sleepMsec }) => {
+                  await new Promise((resolve) => {
+                    setTimeout(resolve, sleepMsec);
+                  });
+                  return {
+                    content: [
+                      {
+                        type: "text",
+                        text: JSON.stringify({
+                          message: "done",
+                        }),
+                      },
+                    ],
+                  };
+                });
 
     if (options.testHeaders) {
-      server.tool(
-        "check_headers",
-        "Check if specific headers were received",
-        { headerName: z.string() },
-        async ({ headerName }, extra) => {
-          // Get headers for this session
-          const sessionId = extra.sessionId || "default";
-          const headers = sessionHeaders[sessionId] || {};
-          return {
-            content: [
-              {
-                type: "text",
-                text: headers[headerName.toLowerCase()] || "NOT_FOUND",
-              },
-            ],
-          };
-        }
-      );
+      server.registerTool("check_headers", { description: "Check if specific headers were received", inputSchema: z.object({ headerName: z.string() }) }, async ({ headerName }, ctx) => {
+                      // Get headers for this session
+                      const sessionId = ctx.sessionId || "default";
+                      const headers = sessionHeaders[sessionId] || {};
+                      return {
+                        content: [
+                          {
+                            type: "text",
+                            text: headers[headerName.toLowerCase()] || "NOT_FOUND",
+                          },
+                        ],
+                      };
+                    });
     }
 
-    server.tool(
-      "audio_tool",
-      "A tool that returns a dummy audio content.",
-      // Input schema: a single string 'input'
-      {
-        input: z.string().describe("Some input string for the audio tool"),
-      },
-      async ({ input }) => {
-        // Static base64 encoded minimal WAV file (1-byte silent audio)
-        // This is a valid WAV file: RIFF header, WAVE format, fmt chunk (PCM, 44100Hz, 1 channel, 16-bit), data chunk (1 byte of 0x00)
-        const base64Audio =
-          "UklGRiQAAABXQVZFZm10IBAAAAABAAEARKwAAIhYAQACABAAAABkYXRhAgAAAAEA";
+    server.registerTool("audio_tool", { description: "A tool that returns a dummy audio content.", inputSchema: z.object({
+                  input: z.string().describe("Some input string for the audio tool"),
+                }) }, async ({ input }) => {
+                  // Static base64 encoded minimal WAV file (1-byte silent audio)
+                  // This is a valid WAV file: RIFF header, WAVE format, fmt chunk (PCM, 44100Hz, 1 channel, 16-bit), data chunk (1 byte of 0x00)
+                  const base64Audio =
+                    "UklGRiQAAABXQVZFZm10IBAAAAABAAEARKwAAIhYAQACABAAAABkYXRhAgAAAAEA";
 
-        return {
-          content: [
-            {
-              type: "text",
-              text: `Audio input was: ${input}, server: ${name}`,
-            },
-            {
-              type: "audio",
-              mimeType: "audio/wav",
-              data: base64Audio,
-            },
-          ],
-        };
-      }
-    );
+                  return {
+                    content: [
+                      {
+                        type: "text",
+                        text: `Audio input was: ${input}, server: ${name}`,
+                      },
+                      {
+                        type: "audio",
+                        mimeType: "audio/wav",
+                        data: base64Audio,
+                      },
+                    ],
+                  };
+                });
 
-    server.tool(
-      "image_tool",
-      "A tool that returns a dummy image and text content.",
-      // Input schema: a single string 'input'
-      {
-        input: z.string().describe("Some input string for the image tool"),
-      },
-      async ({ input }) => {
-        // Static base64 encoded minimal PNG file (1x1 black pixel)
-        const base64Image =
-          "iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mNkYAAAAAYAAjCB0C8AAAAASUVORK5CYII=";
+    server.registerTool("image_tool", { description: "A tool that returns a dummy image and text content.", inputSchema: z.object({
+                  input: z.string().describe("Some input string for the image tool"),
+                }) }, async ({ input }) => {
+                  // Static base64 encoded minimal PNG file (1x1 black pixel)
+                  const base64Image =
+                    "iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mNkYAAAAAYAAjCB0C8AAAAASUVORK5CYII=";
 
-        return {
-          content: [
-            {
-              type: "text",
-              text: `Image input was: ${input}, server: ${name}`,
-            },
-            {
-              type: "image",
-              mimeType: "image/png",
-              data: base64Image,
-            },
-          ],
-        };
-      }
-    );
+                  return {
+                    content: [
+                      {
+                        type: "text",
+                        text: `Image input was: ${input}, server: ${name}`,
+                      },
+                      {
+                        type: "image",
+                        mimeType: "image/png",
+                        data: base64Image,
+                      },
+                    ],
+                  };
+                });
 
-    server.tool(
-      "resource_tool",
-      "A tool that returns a dummy resource and text content.",
-      {
-        input: z.string().describe("Some input string for the resource tool"),
-      },
-      async ({ input }) => {
-        return {
-          content: [
-            {
-              type: "text",
-              text: `Resource input was: ${input}, server: ${name}`,
-            },
-            {
-              type: "resource",
-              resource: {
-                uri: "mem://test.txt",
-                mimeType: "text/plain",
-                text: "This is a test resource.",
-                // No blob, to test text-based resource handling
-              },
-            },
-          ],
-        };
-      }
-    );
+    server.registerTool("resource_tool", { description: "A tool that returns a dummy resource and text content.", inputSchema: z.object({
+                  input: z.string().describe("Some input string for the resource tool"),
+                }) }, async ({ input }) => {
+                  return {
+                    content: [
+                      {
+                        type: "text",
+                        text: `Resource input was: ${input}, server: ${name}`,
+                      },
+                      {
+                        type: "resource",
+                        resource: {
+                          uri: "mem://test.txt",
+                          mimeType: "text/plain",
+                          text: "This is a test resource.",
+                          // No blob, to test text-based resource handling
+                        },
+                      },
+                    ],
+                  };
+                });
 
     // Add a tool that returns structuredContent and _meta
-    server.tool(
-      "structured_tool",
-      "A tool that returns structuredContent and _meta",
-      {
-        input: z.string().describe("Some input string"),
-      },
-      async ({ input }) => {
-        return {
-          content: [
-            {
-              type: "text",
-              text: `Structured input was: ${input}`,
-            },
-          ],
-          structuredContent: {
-            type: "object",
-            data: {
-              result: "success",
-              value: input,
-              timestamp: new Date().toISOString(),
-            },
-          },
-          _meta: {
-            toolVersion: "1.0.0",
-            serverName: name,
-            executionTime: 100,
-          },
-        };
-      }
-    );
+    server.registerTool("structured_tool", { description: "A tool that returns structuredContent and _meta", inputSchema: z.object({
+                  input: z.string().describe("Some input string"),
+                }) }, async ({ input }) => {
+                  return {
+                    content: [
+                      {
+                        type: "text",
+                        text: `Structured input was: ${input}`,
+                      },
+                    ],
+                    structuredContent: {
+                      type: "object",
+                      data: {
+                        result: "success",
+                        value: input,
+                        timestamp: new Date().toISOString(),
+                      },
+                    },
+                    _meta: {
+                      toolVersion: "1.0.0",
+                      serverName: name,
+                      executionTime: 100,
+                    },
+                  };
+                });
 
     // Add resource handlers
     server.registerResource(
@@ -320,7 +278,7 @@ export function createDummyHttpServer(
 
   // Store transports and metadata
   const transports = {
-    streamable: {} as Record<string, StreamableHTTPServerTransport>,
+    streamable: {} as Record<string, NodeStreamableHTTPServerTransport>,
     sse: {} as Record<string, SSEServerTransport>,
   };
 
@@ -342,7 +300,7 @@ export function createDummyHttpServer(
   if (!options.disableStreamableHttp) {
     app.post("/mcp", async (req, res) => {
       const sessionId = req.headers["mcp-session-id"] as string | undefined;
-      let transport: StreamableHTTPServerTransport;
+      let transport: NodeStreamableHTTPServerTransport;
 
       if (options.requireAuth) {
         const auth = req.headers.authorization;
@@ -361,7 +319,7 @@ export function createDummyHttpServer(
         // Capture headers for existing session
         captureHeaders(req, sessionId);
       } else if (!sessionId && isInitializeRequest(req.body)) {
-        transport = new StreamableHTTPServerTransport({
+        transport = new NodeStreamableHTTPServerTransport({
           sessionIdGenerator: () => randomUUID(),
           onsessioninitialized: (newSessionId) => {
             transports.streamable[newSessionId] = transport;
